# swagger_server/models/new_password.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class NewPassword(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, new_password: str=None):  # noqa: E501
        """NewPassword - a model defined in Swagger

        :param new_password: The new_password of this NewPassword.  # noqa: E501
        :type new_password: str
        """
        self.swagger_types = {
            'new_password': str
        }

        self.attribute_map = {
            'new_password': 'new_password'
        }
        self._new_password = new_password

    @classmethod
    def from_dict(cls, dikt) -> 'NewPassword':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The NewPassword of this NewPassword.  # noqa: E501
        :rtype: NewPassword
        """
        return util.deserialize_model(dikt, cls)

    @property
    def new_password(self) -> str:
        """Gets the new_password of this NewPassword.

        New password  # noqa: E501

        :return: The new_password of this NewPassword.
        :rtype: str
        """
        return self._new_password

    @new_password.setter
    def new_password(self, new_password: str):
        """Sets the new_password of this NewPassword.

        New password  # noqa: E501

        :param new_password: The new_password of this NewPassword.
        :type new_password: str
        """

        self._new_password = new_password